requirements.txt
Django==4.2.3
Pillow==10.0.0
stripe==6.0.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
psutil
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.95.0
python-dotenv==0.19.1